root/requirements.txt
alembic==1.12.1
amqp==5.2.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
autobahn==23.6.2
Automat==22.10.0
azure-common==1.1.28
azure-core==1.29.5
azure-identity==1.13.0
azure-mgmt-core==1.4.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.0.1
azure-mgmt-subscription==3.1.1
Babel==2.13.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bidict==0.22.1
billiard==4.2.0
blinker==1.4
boto3==1.28.73
botocore==1.31.73
Brlapi==0.8.3
Brotli==1.1.0
cachetools==5.3.2
celery==5.3.4
certifi==2020.6.20
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
chardet==4.0.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
command-not-found==0.3
constantly==23.10.4
contourpy==1.1.1
cryptography==41.0.5
cupshelpers==1.0
cycler==0.12.1
daphne==4.0.0
defer==1.0.6
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.6
django-compat==1.0.15
django-cors-headers==4.3.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
email-validator==2.1.0.post1
eventlet==0.33.3
fail2ban==0.11.2
filelock==3.13.0
Flask==2.2.5
flask-babel==3.1.0
Flask-Compress==1.14
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.1.2
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fonttools==4.43.1
frozendict==2.3.8
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.61.0
greenlet==1.1.2
gyp==0.1
h11==0.14.0
html5lib==1.1
httpagentparser==1.9.5
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==22.10.0
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
keyring==23.5.0
kiwisolver==1.4.5
kombu==5.3.3
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
ldap3==2.9.1
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.8.1
more-itertools==8.10.0
msal==1.24.1
msal-extensions==1.0.0
msgpack==1.0.7
msrest==0.7.1
multitasking==0.0.11
netifaces==0.11.0
numpy==1.26.1
oauthlib==3.2.0
olefile==0.46
packaging==23.2
pandas==2.1.2
paramiko==3.3.1
passlib==1.7.4
peewee==3.17.0
pexpect==4.8.0
pgadmin4==7.8
Pillow==9.0.1
platformdirs==3.11.0
portalocker==2.8.2
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
psycopg==3.1.12
psycopg-binary==3.1.12
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.11.2
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.3.0
pyotp==2.9.0
pyparsing==2.4.7
pypng==0.20220715.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.4.2
redis==5.0.1
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.7.0
scipy==1.11.3
SecretStorage==3.3.1
service-identity==23.1.0
simple-websocket==1.0.0
six==1.16.0
soupsieve==2.5
speaklater3==1.4
SQLAlchemy==2.0.22
sqlparse==0.4.4
ssh-import-id==5.11
sshtunnel==0.4.0
supervisor==4.2.1
systemd-python==234
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
ua-parser==0.18.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
user-agents==2.2.0
vaderSentiment==3.3.2
vine==5.1.0
virtualenv==20.24.6
wadllib==1.3.6
wcwidth==0.2.9
webencodings==0.5.1
Werkzeug==2.2.3
wsproto==1.2.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
yfinance==0.2.31
zipp==1.0.0
zope.interface==6.1
